refactor(core): extract theme builtins tables and themed() constructor
Move ThemeTokens, frozen BUILTIN_THEMES / LEGACY_BUILTIN_THEMES, and themed()
into theme-builtins.ts so pure registry data is not mixed with resolveTheme /
themeVar. Public ./theme.js re-exports keep package and deep imports stable.
